naming ok pour une des fonctions d autres a faire ?
