+++ server/src/modules/Trip/tripActions.ts
@@ -43,14 +43,5 @@ const browse = async (req: Request, res: Response) => {
 		res.sendStatus(500);
 	}
 };
-const browseAll: RequestHandler = async (req, res, next) => {
-	try {
-		const themeId = Number(req.params.id);
-		const trips = await tripRepository.readAll(themeId);
-		res.json(trips);
-	} catch (err) {
-		next(err);
-	}
-};
 
-export default { add, browseAllByCountry, browse, browseAll };
+export default { add, browseAllByCountry, browse };
